requirements.txt
torch
torchvision
numpy
Pillow
torch
pandas
timm
scikit-learn
opencv-python
pydicom
albumentations
monai